added set_time table. add foreign keys to tables

+++ models/stage_events.js
@@ -21,11 +21,11 @@ module.exports = (sequelize, DataTypes) => {
       autoIncrement: true
     },
     stage_id: {
-      type: DataTypes.SMALLINT,
+      type: DataTypes.INTEGER,
       allowNull: false
     },
     event_id: {
-      type: DataTypes.SMALLINT,
+      type: DataTypes.INTEGER,
       allowNull: false
     }
   }, {
